+++ Projects/Features/SearchFeature/Demo/Sources/AppDelegate.swift
@@ -25,7 +25,7 @@ final class AppDelegate: UIResponder, UIApplicationDelegate {
 //        let component = SongSearchResultViewController(reactor: SongSearchResultReactor("1234"))
 //        let component2 = ListSearchResultViewController(reactor: ListSearchResultReactor("text"))
 
-        let component = SearchOptionViewController(selectedModel: .latest)
+        let component = SearchSortOptionViewController(selectedModel: .latest)
 
         let viewController = Inject.ViewControllerHost(
             UINavigationController(rootViewController: component)
