+++ Projects/Features/SearchFeature/Sources/Components/SearchOptionComponent.swift
@@ -5,6 +5,6 @@ import UIKit
 
 public final class SearchOptionComponent: Component<EmptyDependency> {
     public func makeView(_ sortType: SortType) -> UIViewController {
-        return SearchOptionViewController(selectedModel: sortType)
+        return SearchSortOptionViewController(selectedModel: sortType)
     }
 }
