+++ Projects/Features/SearchFeature/Sources/CompositionalLayout/Layout/SearchOptionCollectionViewLayout.swift
@@ -1,0 +1,13 @@
+//
+//  SearchOptionCollectionViewLayout.swift
+//  SearchFeature
+//
+//  Created by yongbeomkwak on 6/18/24.
+//  Copyright © 2024 yongbeomkwak. All rights reserved.
+//
+
+import UIKit
+
+class SearchOptionCollectionViewLayout: UICollectionViewLayout {
+
+}
