+++ Projects/Features/SearchFeature/Sources/View/SearchFilterCell.swift
@@ -1,0 +1,13 @@
+//
+//  SearchFilterCell.swift
+//  SearchFeature
+//
+//  Created by yongbeomkwak on 6/18/24.
+//  Copyright © 2024 yongbeomkwak. All rights reserved.
+//
+
+import UIKit
+
+class SearchFilterCell: UICollectionViewCell {
+    
+}
